Rafactor: Move interfaces between folders, delete unused files
