+++ Filc/Services/Interfaces/EntityBasedInterfaces/ISchoolAdminService.cs
@@ -1,0 +1,9 @@
+﻿using EFDataAccessLibrary.Models;
+
+namespace Filc.Services.New_intefaces
+{
+    public interface ISchoolAdminService
+    {
+        public void AddSchoolAdmin(int schoolId, SchoolAdmin schoolAdmin);
+    }
+}
