+++ Filc/Services/Interfaces/EntityBasedInterfaces/ISchoolService.cs
@@ -1,12 +1,13 @@
 ﻿using EFDataAccessLibrary.Models;
 
-namespace Filc.Services.Interfaces
+namespace Filc.Services.New_intefaces.EntityBasedInterfaces
 {
-    public interface IGovernmentService
+    public interface ISchoolService
     {
         public void AddSchool(School school);
         public void RemoveSchool(int id);
+        public void UpdateSchool(School school);
+        public School GetSchool(int id);
         public List<School> ReturnAllSchools();
-        public void AddSchoolAdmin(int schoolId, SchoolAdmin schoolAdmin);
     }
 }
