--- Filc/Services/Interfaces/ITeacherService.cs
@@ -1,10 +1,0 @@
-﻿using EFDataAccessLibrary.Models;
-
-namespace Filc.Services.Interfaces
-{
-    public interface ITeacherService
-    {
-        public void AddMark(Mark mark);
-        public void RegisterAbsence();
-    }
-}
