--- Filc/Services/New intefaces/EntityBasedInterfaces/ISchoolAdminService.cs
@@ -1,6 +1,0 @@
-﻿namespace Filc.Services.New_intefaces
-{
-    public class ISchoolAdminService
-    {
-    }
-}
